fix(ci): remove integration test from build script test job
The build-script-test job was trying to run ./test_integration but this
binary is not built in this job. The integration test is already properly
tested in the cmake-build and autotools-build jobs where the test binary
is actually compiled.

The build-script-test job should focus on testing the build script
functionality, not running integration tests.

Fixes error:
./test_integration: No such file or directory